fix(hasRoleOrRestrict): usage w/ get method returns correct result (#28)
previously the hasRoleOrRestrict hook failed to return data in the expected format when used on a
Service#get method due to incorrect handling of the result of the internal call to Service#find. as
Service#find returns an object with a "data" property that is always an array, we can safely check
the nested value and correctly determine if the pseudo-get method call was successful.

any application code hacking around the issue by knowingly pulling the nested "data" prop from the
result will now break.